Implemented the majority of the remaining bits of serialization

+++ scala/src/org/naerbnic/wrapt/primitive/LitIndexType.scala
@@ -3,7 +3,7 @@ package org.naerbnic.wrapt.primitive
 import org.naerbnic.wrapt.util.LongBits._
 import org.naerbnic.wrapt.BasicValue._
 
-sealed abstract class LitIndexType(val code: Int) {
+sealed abstract class LitIndexType(val code: Long) {
   def getValue(data: Long): PrimValue
 }
